Drop `Send` in `Command` and `Executor` on Wasm
